put mapped guid stuff into new files
